requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-filter==23.3
drf_yasg==1.21.7
django-cors-headers==4.2.0
psycopg2==2.9.7
Pillow==10.0.1
phonenumbers==8.13.21
pandas==2.1.1
num2words==0.5.12
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
uvicorn==0.23.2
django-phonenumber-field==7.1.0
psycopg2-binary==2.9.7
django-tinymce==3.6.1
